move  addversion into separate mixin
